% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{print.ptvalue}
\alias{print.ptvalue}
\title{Title}
\usage{
\method{print}{ptvalue}(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Other args}
}
\value{
A ptvalue
}
\description{
Title
}
\examples{
#Todo
}
