% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{abs_sign}
\alias{abs_sign}
\title{Find 'absolute' ptvalue}
\usage{
abs_sign(x = double(), sign = "times")
}
\arguments{
\item{x}{A vector of class \code{ptvalue}}

\item{sign}{Either 'times' or 'div'. Default to 'times'.}
}
\value{
A vector of class \code{ptvalue} with absolute ptvalue
}
\description{
Find 'absolute' ptvalue
}
\examples{
x <- c(0.5, 1.4, 2)
abs_sign(x)
}
