% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{invert_sign}
\alias{invert_sign}
\title{Invert ptvalue sign}
\usage{
invert_sign(x = double())
}
\arguments{
\item{x}{A vector of class \code{ptvalue}}
}
\value{
A vector of class \code{ptvalue} with inverted sign
}
\description{
Invert ptvalue sign
}
\examples{
x <- c(0.5, 1.4, 2)
invert_sign(x)
}
