% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptvalue.R
\name{times}
\alias{times}
\alias{div}
\title{Create times or div vector of class \strong{ptvalue}.}
\usage{
times(x = double())

div(x = double())
}
\arguments{
\item{x}{A numeric vector. Values must be greater or equal than 1.}
}
\value{
A numeric vector of class \strong{ptvalue} that represent precision teaching mesures.
}
\description{
\code{times()} and \code{div()} are convenient and stricter functions
for creating growing or decaying precision teaching values with values
with values greater or equal than 1 (or otherwise raise an error).
}
\examples{
times(c(1, 2, 4))
div(c(1, 2, 4))
}
